pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "python-slack-agents"
version = "0.5.0"
description = "A Python framework for deploying AI agents as Slack bots"
authors = [{name = "Eric Pichon", email = "epichon@comparenetworks.com"}]
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.12"
keywords = ["slack", "ai", "agents", "chatbot", "llm", "mcp", "anthropic", "openai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Chat",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "slack-bolt[async]>=1.20,<2",
    "slack-sdk[socket-mode-handlers]>=3.33,<4",
    "aiohttp>=3.9,<4",
    "mcp>=1.0,<2",
    "anthropic>=0.40,<1",
    "openai>=1.50,<2",
    "asyncpg>=0.30,<1",
    "aiosqlite>=0.20,<1",
    "opentelemetry-sdk>=1.20,<2",
    "opentelemetry-exporter-otlp-proto-http>=1.20,<2",
    "pydantic>=2.0,<3",
    "pyyaml>=6.0,<7",
    "python-dotenv>=1.0,<2",
    "httpx>=0.27,<1",
    "pymupdf>=1.24,<2",
    "pymupdf4llm>=0.3,<1",
    "python-docx>=1.1,<2",
    "openpyxl>=3.1,<4",
    "fpdf2>=2.8,<3",
    "python-pptx>=1.0,<2",
]

[project.urls]
Homepage = "https://github.com/CompareNetworks/python-slack-agents"
Repository = "https://github.com/CompareNetworks/python-slack-agents"
Documentation = "https://github.com/CompareNetworks/python-slack-agents/tree/main/docs"
Issues = "https://github.com/CompareNetworks/python-slack-agents/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-asyncio>=0.24,<1",
    "ruff>=0.8,<1",
    "pre-commit>=4.0,<5",
]

[project.scripts]
slack-agents = "slack_agents.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/slack_agents"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
